use pytest temp factory

+++ tests/python_tests/samples/conftest.py
@@ -1,6 +1,5 @@
 import subprocess # nosec B404
 import os
-import tempfile
 import pytest
 import shutil
 import logging
@@ -55,35 +54,49 @@ TEST_FILES = {
     "adapter_model.safetensors": "https://huggingface.co/smangrul/tinyllama_lora_sql/resolve/main/adapter_model.safetensors"
 }
 
-OV_CACHE = os.environ.get("OV_CACHE", tempfile.mkdtemp())
-MODELS_DIR = os.path.join(OV_CACHE, "test_models")
-TEST_DATA = os.path.join(OV_CACHE, "test_data")
-HF_CACHE = os.environ.get("HF_HOME", os.path.join(OV_CACHE, "hf_cache"))
-
 SAMPLES_PY_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../samples/python"))
 SAMPLES_CPP_DIR = os.environ.get("SAMPLES_CPP_DIR", os.getcwd())
 
+def pytest_addoption(parser):
+    parser.addoption(
+        "--cleanup-cache", action="store_true", default=False, help="Cleanup cache after tests"
+    )
+
 @pytest.fixture(scope="session", autouse=True)
-def setup_and_teardown(request):
+def setup_and_teardown(request, tmp_path_factory):
     """Fixture to set up and tear down the temporary directories."""
-    logger.info(f"Creating directories: {MODELS_DIR} and {TEST_DATA}")
-    os.makedirs(MODELS_DIR, exist_ok=True)
-    os.makedirs(TEST_DATA, exist_ok=True)
+    
+    ov_cache = os.environ.get("OV_CACHE", tmp_path_factory.mktemp("ov_cache"))
+    models_dir = os.path.join(ov_cache, "test_models")
+    test_data = os.path.join(ov_cache, "test_data")
+    
+    logger.info(f"Creating directories: {models_dir} and {test_data}")
+    os.makedirs(models_dir, exist_ok=True)
+    os.makedirs(test_data, exist_ok=True)
+    
+    request.config.cache.set("OV_CACHE", str(ov_cache))
+    request.config.cache.set("MODELS_DIR", str(models_dir))
+    request.config.cache.set("TEST_DATA", str(test_data))
+    
     yield
-    if not os.environ.get("OV_CACHE"):
-        logger.info(f"Removing temporary directory: {OV_CACHE}")
-        shutil.rmtree(OV_CACHE)
-    else:
-        logger.info(f"Skipping cleanup of temporary directory: {OV_CACHE}")
+    
+    if request.config.getoption("--cleanup-cache"):
+        if os.path.exists(ov_cache):
+            logger.info(f"Removing temporary directory: {ov_cache}")
+            shutil.rmtree(ov_cache)
+        else:
+            logger.info(f"Skipping cleanup of temporary directory: {ov_cache}")
 
 @pytest.fixture(scope="session")
 def convert_model(request):
     """Fixture to convert the model once for the session."""
+    models_cache = request.config.cache.get("MODELS_DIR", None)
     model_id = request.param
     model_name = MODELS[model_id].get("name")
-    model_path = os.path.join(MODELS_DIR, model_name)
+    model_cache = os.path.join(models_cache, model_id)
+    model_path = os.path.join(model_cache, model_name)
     model_args = MODELS[model_id].get("convert_args", [])
-    model_hf_cache = os.path.join(HF_CACHE, model_id)
+    model_hf_cache = os.environ.get("HF_HOME", os.path.join(model_cache, "hf_cache"))
     logger.info(f"Preparing model: {model_name}")
     # Convert the model if not already converted
     if not os.path.exists(model_path):
@@ -111,26 +124,22 @@ def convert_model(request):
                     raise e
             
     yield model_path
+    
     # Cleanup the model after tests
-    if not os.environ.get("OV_CACHE"):
-        if os.path.exists(model_path):
-            logger.info(f"Removing converted model: {model_path}")
-            shutil.rmtree(model_path)
-    else:
-        logger.info(f"The model {model_path} will be removed when all the tests are finished")
-        
-    # Cleanup the model cache after test
-    if not os.environ.get("HF_HOME"):
-        if os.path.exists(model_hf_cache):
-            logger.info(f"Removing cached model: {model_hf_cache}")
-            shutil.rmtree(model_hf_cache)
+    if request.config.getoption("--cleanup-cache"):
+        if os.path.exists(model_cache):
+            logger.info(f"Removing converted model: {model_cache}")
+            shutil.rmtree(model_cache)
 
 @pytest.fixture(scope="session")
 def download_test_content(request):
     """Download the test content from the given URL and return the file path."""
+    
+    test_data = request.config.cache.get("TEST_DATA", None)
+    
     file_url = request.param
     file_name = os.path.basename(file_url)
-    file_path = os.path.join(TEST_DATA, file_name)
+    file_path = os.path.join(test_data, file_name)
     if not os.path.exists(file_path):
         logger.info(f"Downloading test content from {file_url}...")
         result = subprocess.run(
@@ -143,9 +152,10 @@ def download_test_content(request):
         logger.info(f"Test content already exists at {file_path}")
     yield file_path
     # Cleanup the test content after tests
-    if os.path.exists(file_path):
-        logger.info(f"Removing test content: {file_path}")
-        os.remove(file_path)
+    if request.config.getoption("--cleanup-cache"):
+        if os.path.exists(file_path):
+            logger.info(f"Removing test content: {file_path}")
+            os.remove(file_path)
 
 
 @pytest.fixture(scope="module", autouse=True)
